Cleaned up files and committed words.json
