+++ api/src/main.rs
@@ -1,0 +1,17 @@
+use actix_web::{
+    delete, get, post, put, web, App, HttpRequest, HttpResponse, HttpServer, Responder,
+    ResponseError,
+};
+
+#[get("/hello")]
+async fn hello() -> impl Responder {
+    HttpResponse::Ok().body("Hello, world")
+}
+
+#[actix_web::main]
+async fn main() -> std::io::Result<()> {
+    HttpServer::new(move || App::new().service(hello))
+        .bind(("127.0.0.1", 8000))?
+        .run()
+        .await
+}
